[scripts] Resolve "Project cannot be opened because the project file cannot be parsed." errors when using ttmodule to add Three20 to projects when Three20 itself is contained in a folder path with spaces.
We weren't properly quoting path file names for the Three20 xcodeproj
files, causing Xcode to fail to parse the pbxproj file format when
Three20 was contained in a folder with spaces. This change simply
adds quotes around the path when adding it. It also checks the existence
of a quoted/unquoted path in order to stay backwards compatible.